Bug Fix: NullPointerException when marshalling / unmarshalling using no-args constructor.
Cause: Programming fault. Was not checking for null.

Bug Fix: Property's content may be empty only when using this element as a property NAME. In all other cases, content must be a Date, hence not null. As a side effect, the implementation is simpler and more performance.

Cause: Misunderstanding of the spec.

Bug Fix: equal() returned false for actually equal object.

Cause: False assumption. Object.equals() does not check any members, but solely checks for identity!